streamlit/requirements.txt
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.1.0
reportlab==4.2.2
Requests==2.32.3
streamlit==1.36.0
